gather odesolver tests and testing api
