+++ Component/ODESolver/CMakeLists.txt
@@ -29,3 +29,11 @@ sofa_create_package_with_targets(
     INCLUDE_SOURCE_DIR "src"
     INCLUDE_INSTALL_DIR "${PROJECT_NAME}"
 )
+
+cmake_dependent_option(SOFANG_ENABLE_TESTING_SOFA_COMPONENT_ODESOLVER "Compile the automatic tests" ON "SOFA_BUILD_TESTS OR NOT DEFINED SOFA_BUILD_TESTS" OFF)
+if(SOFANG_ENABLE_TESTING_SOFA_COMPONENT_ODESOLVER)
+    # add testing framework
+    add_subdirectory(Testing)
+    # add tests
+    add_subdirectory(tests)
+endif()
