+++ Component/ODESolver/Testing/src/sofa/component/odesolver/testing/empty.cpp
@@ -1,0 +1,3 @@
+#include <sofa/config.h>
+
+SOFA_EXPORT_DYNAMIC_LIBRARY void init(){}
