+++ Component/ODESolver/tests/VariationalSymplecticExplicitSolverDynamic_test.cpp
@@ -19,13 +19,12 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-
 #include <sofa/testing/BaseSimulationTest.h>
 using sofa::testing::BaseSimulationTest;
 #include <sofa/testing/NumericTest.h>
 using sofa::testing::NumericTest;
 
-#include <SofaExplicitOdeSolver_test/MassSpringSystemCreation.h>
+#include <sofa/component/odesolver/testing/MassSpringSystemCreation.h>
 
 #include <SceneCreator/SceneCreator.h>
 
@@ -56,36 +55,33 @@ Test the dynamic behavior of solver: study a mass-spring system under gravity in
 The movement follows the equation:
 x(t)= A cos(wt + phi) with w the pulsation w=sqrt(K/M), K the stiffness, M the mass and phi the phase.
 In this test: x(t=0)= 1 and v(t=0)=0 and K = spring stiffness and phi = 0 of material thus x(t)= cos(wt)
-This tests generates the discrete mass position obtained with variational symplectic solver with different parameter values (K,M,h).
+This tests generates the discrete mass position obtained with variational symplectic explicit solver with different parameter values (K,M,h).
 Then it compares the effective mass position to the computed mass position every time step.
 */
 
 template <typename _DataTypes>
-struct VariationalSymplecticImplicitSolverDynamic_test : public BaseSimulationTest
+struct VariationalSymplecticExplicitSolverDynamic_test : public BaseSimulationTest
 {
     typedef _DataTypes DataTypes;
     typedef typename DataTypes::Coord Coord;
+    typedef typename DataTypes::Real Real;
 
     typedef container::MechanicalObject<DataTypes> MechanicalObject;
     typedef component::odesolver::VariationalSymplecticSolver VariationalSymplecticSolver;
     typedef component::linearsolver::CGLinearSolver<component::linearsolver::GraphScatteredMatrix, component::linearsolver::GraphScatteredVector> CGLinearSolver;
+
 
     /// Root of the scene graph
     simulation::Node::SPtr root;      
     /// Tested simulation
     simulation::Simulation* simulation;  
     /// Position and velocity array
-    type::vector<double> positionsArray;
-    type::vector<double> velocitiesArray;
-    type::vector<double> energiesArray;
-    // solver
-    VariationalSymplecticSolver::SPtr variationalSolver;
-
-    // totalEnergy
-    double totalEnergy;
-
+    type::vector<Real> positionsArray;
+    type::vector<Real> velocitiesArray;
+    type::vector<Real> energiesArray;
+    
     /// Create the context for the scene
-    void createScene(double K, double m, double l0, double rm=0, double rk=0)
+    void createScene(double K, double m, double l0, double rm=0)
     {
         // Init simulation
         sofa::simulation::setSimulation(simulation = new sofa::simulation::graph::DAGSimulation());
@@ -95,17 +91,15 @@ struct VariationalSymplecticImplicitSolverDynamic_test : public BaseSimulationTe
         root->setGravity(Coord(0,-10,0));
 
         // Solver
-        variationalSolver = addNew<VariationalSymplecticSolver> (root);
-        variationalSolver->f_rayleighStiffness.setValue(rk);
+        VariationalSymplecticSolver::SPtr variationalSolver = addNew<VariationalSymplecticSolver> (root);
+        // Explicit solver
+        variationalSolver->f_explicit.setValue("true");
         variationalSolver->f_rayleighMass.setValue(rm);
-        variationalSolver->f_computeHamiltonian.setValue(1);
-        variationalSolver->f_newtonError.setValue(1e-12);//1e-18
-        variationalSolver->f_newtonSteps.setValue(4);//7
 
         CGLinearSolver::SPtr cgLinearSolver = addNew<CGLinearSolver> (root);
         cgLinearSolver->d_maxIter.setValue(3000);
-        cgLinearSolver->d_tolerance.setValue(1e-12);
-        cgLinearSolver->d_smallDenominatorThreshold.setValue(1e-12);
+        cgLinearSolver->d_tolerance.setValue(1e-9);
+        cgLinearSolver->d_smallDenominatorThreshold.setValue(1e-9);
 
         // Set initial positions and velocities of fixed point and mass
         MechanicalObject3::VecCoord xFixed(1);
@@ -115,7 +109,7 @@ struct VariationalSymplecticImplicitSolverDynamic_test : public BaseSimulationTe
         MechanicalObject3::VecCoord xMass(1);
         MechanicalObject3::DataTypes::set( xMass[0], 0., 1.,0.);
         MechanicalObject3::VecDeriv vMass(1);
-        MechanicalObject3::DataTypes::set( vMass[0], 0., 0., 0.);
+        MechanicalObject3::DataTypes::set( vFixed[0], 0.,0.,0.);
 
         // Mass spring system
         root = sofa::createMassSpringSystem<DataTypes>(
@@ -127,11 +121,11 @@ struct VariationalSymplecticImplicitSolverDynamic_test : public BaseSimulationTe
                 vFixed, // Initial velocity of fixed point
                 xMass,  // Initial position of mass
                 vMass); // Initial velocity of mass
+
     }
 
-
     /// Generate discrete mass position values with variational sympletic implicit solver
-    void generateDiscreteMassPositions (double h, double K, double m, double z0, double v0,double g, double finalTime, double rm,double rk)
+    void generateDiscreteMassPositions (double h, double K, double m, double z0, double v0,double g, double finalTime, double rm)
     {
         int size = 0 ;
 
@@ -153,25 +147,18 @@ struct VariationalSymplecticImplicitSolverDynamic_test : public BaseSimulationTe
         // First position is z0
         positionsArray.push_back(double(z0));
 
-        // Compute totalEnergy
-        totalEnergy = m*g*z0; // energy at initial time
-
-        // Set constants
-        double denominator = 4*m+h*h*K+4*h*(rm*m + rk*K);//4*h*(-rk*K+rm*m);
-        double constant = -h*K;
-
         // Compute velocities, energies and positions
         for(int i=1;i< size+1; i++)
         {
-            velocitiesArray.push_back(2*(-m*g*h+constant*(positionsArray[i-1]-z0)+2*energiesArray[i-1])/denominator);
-            energiesArray.push_back(m*velocitiesArray[i]+h*(-K*(positionsArray[i-1]- z0+velocitiesArray[i]*h/2) -m*g)/2);
+            velocitiesArray.push_back((energiesArray[i-1]+h*(-K*(positionsArray[i-1]-z0)-m*g-rm*m*velocitiesArray[i-1]))/m);
+            energiesArray.push_back(m*velocitiesArray[i]);
             positionsArray.push_back(positionsArray[i-1]+h*velocitiesArray[i]);
         }
 
     }
 
     /// After simulation compare the positions of points to the theoretical positions.
-    bool compareSimulatedToTheoreticalPositions( double h, double tolerancePosition, double toleranceEnergy = 1e-13, double checkEnergyConservation=false)
+    bool compareSimulatedToTheoreticalPositions(double tolerance,double h)
     {
         int i = 0;
         // Init simulation
@@ -184,14 +171,14 @@ struct VariationalSymplecticImplicitSolverDynamic_test : public BaseSimulationTe
 
         // Animate
         do
-        {
+        {              
             // Record the mass position
             Coord p0=dofs.get()->read(sofa::core::ConstVecCoordId::position())->getValue()[0];
 
             double absoluteError = fabs(p0[1]-positionsArray[i]);
 
             // Compare mass position to the theoretical position
-            if( absoluteError > tolerancePosition )
+            if( absoluteError > tolerance )
             {
                 ADD_FAILURE() << "Position of mass at time " << time << " is wrong: "  << std::endl
                     <<" expected Position is " << positionsArray[i] << std::endl
@@ -203,18 +190,6 @@ struct VariationalSymplecticImplicitSolverDynamic_test : public BaseSimulationTe
             //Animate
             sofa::simulation::getSimulation()->animate(root.get(),h);
             time = root->getTime();
-
-
-            // Check if hamiltonian energy is constant when there is no damping
-            if(checkEnergyConservation && fabs(variationalSolver->f_hamiltonianEnergy.getValue() -totalEnergy) > toleranceEnergy )
-            {
-                ADD_FAILURE() << "Hamiltonian energy at time " << time << " is wrong: "  << std::endl
-                    <<" expected Energy is " << totalEnergy << std::endl
-                    <<" actual Energy is   " << variationalSolver->f_hamiltonianEnergy.getValue() << std::endl
-                    << "absolute error     = " << fabs(variationalSolver->f_hamiltonianEnergy.getValue() -totalEnergy) << std::endl;
-                return false;
-            }
-
             // Iterate
             i++;
         }
@@ -231,41 +206,30 @@ typedef Types<
 > DataTypes; // the types to instanciate.
 
 // Test suite for all the instanciations
-TYPED_TEST_SUITE(VariationalSymplecticImplicitSolverDynamic_test, DataTypes);
+TYPED_TEST_SUITE(VariationalSymplecticExplicitSolverDynamic_test, DataTypes);
 
-// Test case: h=0.1 k=100 m =1 rm=0.1 rk=0.1
-TYPED_TEST( VariationalSymplecticImplicitSolverDynamic_test , variationalSymplecticImplicitSolverDynamicTest_high_dt_without_damping)
+// Test case: h=0.001
+TYPED_TEST( VariationalSymplecticExplicitSolverDynamic_test , variationalSymplecticExplicitSolverDynamicTest_small_dt_without_damping)
 {
-   this->createScene(100,1,1,0,0); // k,m,l0
-   // rm = 0.1  and rk=0
-   this->generateDiscreteMassPositions (0.01, 100, 1, 1, 0,10, 2, 0 ,0);
-   // With incremental potential energy
-   this-> compareSimulatedToTheoreticalPositions(0.01,9e-15,8e-15,true);
+   this->createScene(100,10,1,0); // k,m,l0
+   this->generateDiscreteMassPositions (0.001, 100, 10, 1, 0, 10, 2, 0);
+   this-> compareSimulatedToTheoreticalPositions(1.9e-14,0.001);
 }
 
-// Test case: h=0.01 K=100 m = 10 rm=0.1 rk=0
-TYPED_TEST( VariationalSymplecticImplicitSolverDynamic_test , variationalSymplecticImplicitSolverDynamicTest_medium_dt_with_rayleigh_mass)
+// Test case: h=0.01 k=1000 m =10 rm=0.1
+TYPED_TEST( VariationalSymplecticExplicitSolverDynamic_test , variationalSymplecticExplicitSolverDynamicTest_medium_dt_with_damping)
 {
-   this->createScene(100,10,1,0.1,0); // k,m,l0
-   this->generateDiscreteMassPositions (0.01, 100, 10, 1, 0,10, 2, 0.1,0);
-   this-> compareSimulatedToTheoreticalPositions(0.01, 7e-15);
+     this->createScene(1000,10,1,0.1); // k,m,l0
+     this->generateDiscreteMassPositions (0.01, 1000, 10, 1, 0, 10, 2, 0.1);
+     this-> compareSimulatedToTheoreticalPositions(2.3e-16,0.01);
 }
 
-// Test case: h=0.01 k=100 m=10 rm=0 rk=0.1
-TYPED_TEST( VariationalSymplecticImplicitSolverDynamic_test , variationalSymplecticImplicitSolverDynamicTest_medium_dt_with_rayleigh_stiffness)
+// Test case: h=0.1 K=10 m=10 rm=0.1
+TYPED_TEST( VariationalSymplecticExplicitSolverDynamic_test , variationalSymplecticExplicitSolverDynamicTest_high_dt_with_damping)
 {
-   this->createScene(100,10,1,0,0.1); // k,m,l0
-   this->generateDiscreteMassPositions (0.01, 100, 10, 1, 0,10, 2, 0, 0.1);
-   this-> compareSimulatedToTheoreticalPositions(0.01, 3.3e-15);
-}
-
-// Test case: h=0.001 k=100 m =10 rm=0.1 rk=0.1
-TYPED_TEST( VariationalSymplecticImplicitSolverDynamic_test , variationalSymplecticImplicitSolverDynamicTest_small_dt_with_damping)
-{
-   this->createScene(100,10,1,0.1,0.1); // k,m,l0
-   // rm = 0.1  and rk=0
-   this->generateDiscreteMassPositions (0.001, 100, 10, 1, 0,10, 2, 0.1 ,0.1);
-   this-> compareSimulatedToTheoreticalPositions(0.001, 6.3e-15);
+   this->createScene(10,10,1,0.1); // k,m,l0
+   this->generateDiscreteMassPositions (0.1, 10, 10, 1, 0, 10, 2, 0.1);
+   this-> compareSimulatedToTheoreticalPositions(1.8e-15,0.1);
 }
 
 } // namespace sofa
