+++ SofaKernel/modules/SofaImplicitOdeSolver/SofaImplicitOdeSolver_test/CMakeLists.txt
@@ -2,14 +2,8 @@ cmake_minimum_required(VERSION 3.12)
 
 project(SofaImplicitOdeSolver_test)
 
-set(SOURCE_FILES
-    loadPlugins.cpp
-    EulerImplicitSolverStatic_test.cpp
-    EulerImplicitSolverDynamic_test.cpp
-    SpringSolverDynamic_test.cpp
-    StaticSolver_test.cpp)
+set(SOURCE_FILES)
     
-add_definitions("-DSOFAIMPLICITODESOLVER_TEST_SCENES_DIR=\"${CMAKE_CURRENT_SOURCE_DIR}/scenes\"")
 
 add_executable(${PROJECT_NAME} ${SOURCE_FILES})
