+++ modules/SofaMiscSolver/CMakeLists.txt
@@ -30,11 +30,3 @@ sofa_create_package_with_targets(
     INCLUDE_INSTALL_DIR "${PROJECT_NAME}"
     RELOCATABLE "plugins"
     )
-
-# Tests
-# If SOFA_BUILD_TESTS exists and is OFF, then these tests will be auto-disabled
-cmake_dependent_option(SOFAMISCSOLVER_BUILD_TESTS "Compile the automatic tests" ON "SOFA_BUILD_TESTS OR NOT DEFINED SOFA_BUILD_TESTS" OFF)
-if(SOFAMISCSOLVER_BUILD_TESTS)
-    enable_testing()
-    add_subdirectory(SofaMiscSolver_test)
-endif()
